+++ src/main/java/io/kontur/layers/dto/CollectionsSearchDto.java
@@ -1,40 +1,26 @@
 package io.kontur.layers.dto;
 
 import io.kontur.layers.dto.validation.ValidGeoJSON;
+import io.swagger.v3.oas.annotations.media.Schema;
+import lombok.Getter;
+import lombok.Setter;
 import org.wololo.geojson.Geometry;
 
 import javax.validation.constraints.Min;
 
+@Getter
+@Setter
 public class CollectionsSearchDto {
 
     @ValidGeoJSON
     private Geometry geometry;
+    @Schema(defaultValue = "false")
+    private Boolean ownedByUser;
     @Min(1)
+    @Schema(defaultValue = "10")
     private Integer limit = 10;
     @Min(0)
+    @Schema(defaultValue = "0")
     private Integer offset = 0;
 
-    public Geometry getGeometry() {
-        return geometry;
-    }
-
-    public void setGeometry(Geometry geometry) {
-        this.geometry = geometry;
-    }
-
-    public Integer getLimit() {
-        return limit;
-    }
-
-    public void setLimit(Integer limit) {
-        this.limit = limit;
-    }
-
-    public Integer getOffset() {
-        return offset;
-    }
-
-    public void setOffset(Integer offset) {
-        this.offset = offset;
-    }
 }