+++ src/main/java/io/kontur/layers/repository/LayerMapper.java
@@ -13,17 +13,19 @@ public interface LayerMapper {
     Optional<Layer> getOwnedLayer(String publicId, String userName);
 
     default Optional<Layer> getLayer(String publicId, String userName) {
-        return getLayers(null, userName,1, 0, publicId).stream().findFirst();
+        return getLayers(null, userName,1, 0, false, publicId).stream().findFirst();
     }
 
     Integer getLayersTotal(@Param("geometry") String geometry,
-                          @Param("userName") String userName,
-                          @Param("publicIds") String... publicIds);
+                           @Param("userName") String userName,
+                           @Param("ownedByUser") Boolean ownedByUser,
+                           @Param("publicIds") String... publicIds);
 
     List<Layer> getLayers(@Param("geometry") String geometry,
                           @Param("userName") String userName,
                           @Param("limit") Integer limit,
                           @Param("offset") Integer offset,
+                          @Param("ownedByUser") Boolean ownedByUser,
                           @Param("publicIds") String... publicIds);
 
     Layer insertLayer(Layer layer);